Refactor mod and file handling for improved texture and model management
- Added g_TexModNum to track texture modifications across different mods.
- Enhanced readModelTextures to dynamically allocate texture memory and prevent out-of-bounds access.
- Updated fsInit to support multiple mod directories, specifically for mod_fojo and mod_gex_characters.
- Modified modConfigParseHeadOrBodyEntry to handle additional configuration fields and improve error handling.
- Implemented alternate ROM data source handling in romdataFileLoad for better file management.
- Adjusted file handling functions to support expanded file IDs and ensure proper memory allocation.
- Changed body and head file number types from u16 to u32 for better compatibility with mod expansions.
- Updated modeldef functions to handle new file ID types and improve texture remapping.
- Enhanced error logging for file loading and configuration parsing to aid in debugging.

+++ src/game/menu.c
@@ -1932,8 +1932,8 @@ Gfx *menuRenderModel(Gfx *gdl, struct menumodel *menumodel, s32 modeltype)
 	s32 totalfilelen;
 	struct texpool texpool;
 	s32 bodyfilelen2;
-	u16 bodyfilenum;
-	u16 headfilenum;
+	u32 bodyfilenum;
+	u32 headfilenum;
 	s32 bodynum;
 	s32 headnum;
